Move AssetExtensionHelper one level up
